- Added video encoder for automatically making videos of the simulation. - Next is adding proper anti-aliasing to output video and perhaps simulation recreation at a certain time.

+++ src/main/Utils.java
@@ -6,9 +6,9 @@ public class Utils {
      * @param fps - The FPS value to convert to MSPF.
      * @returns the MSPF value.
      */
-    public static double FPStoMSPF(double fps){
+    public static long FPStoMSPF(double fps){
         // frames/second TO milliseconds/frame
-        return (1/fps) * 1000;
+        return (long) ((1/fps) * 1000);
     }
 
     /**
